backend/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
DateTime==4.3
debugpy==1.3.0
Django==3.2.6
django-admin-honeypot==1.1.0
django-background-task==0.1.8
django-cors-headers==3.8.0
django-filter==21.1
django-otp==1.0.6
djangorestframework==3.12.4
jsonfield==3.1.0
jsonschema==3.2.0
numpy==1.19.5
pandas==1.3.0
regex==2021.7.6
requests==2.25.1
yfinance==0.1.67
django-admin-honeypot
django-otp==1.0.6
APScheduler==3.0.0
django-apscheduler==0.6.0
